Trying to fix stack issue
